Some additional Windows nightly build fixes: - restore the original mingw/msys link, even when the build script abort (due to a die command) - only restore the mingw/msys link if it was saved before - weekly builds should be on Monday, not Tuesday - upload tag rebuilds to a directory named 'releases' instead of 'tags'
BP

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@20456 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ packaging/win32/functions.sh
@@ -121,6 +121,7 @@ function die() {
     echo
     [ "$*" ] && echo "!!! $* !!!"
     echo "!!! ABORTING !!!"
+    restore_msys
     exit -1
 }
 
@@ -214,9 +215,11 @@ function configure_msys() {
 
 function restore_msys() {
     SUFFIX=$1
-    echo "resetting msys to use original mingw."
-    rm /etc/fstab
-    mv /etc/fstab.$SUFFIX /etc/fstab
+    if [ -f /ect/fstab.$SUFFIX ]; then
+      echo "resetting msys to use original mingw."
+      rm /etc/fstab
+      mv /etc/fstab.$SUFFIX /etc/fstab
+    fi
 }
 
 ### Local Variables: ***
